Reduce the risk of id clashes
When rendering over 10 charts on a page, I got a periodic clash in the IDs.  This results in the wrong data being loaded into one chart and the other chart not to render.  I have patched this locally to use the logic:

1 . Get the md5 hash of the seed string (in this case, the inner content of the expandable)
2. Split that into an array
3. Shuffle the array
4. Convert the array back into a string 

This gives you a unique id like `584abfca542dc4dbc81eccfcd6e4fd8c`.  This is much less likely to clash (but not impossible).

I hope this is useful to someone else.